backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
sqlalchemy>=2.0.35
pymysql==1.1.0
cryptography==41.0.7
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
pydantic>=2.9.0
pydantic-settings>=2.6.0
email-validator>=2.0.0
python-dotenv==1.0.0
qrcode[pil]>=7.4.2
Pillow>=10.0.0

